requirements.txt
lapx
torch
torchvision
torchaudio
pydantic
opencv-contrib-python-headless
openpyxl
geopandas
folium
leafmap
streamlit
streamlit-folium
ultralytics
streamlit_extras
extra-streamlit-components
pytubefix
yt-dlp
easyocr
streamlit-aggrid
st-btn-group
streamlit-autorefresh
streamlit-webrtc